requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai

google-adk
google-auth
google-auth-oauthlib
requests
